base/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-autoslug==1.9.9
django-cities-light==3.10.1
django-countries==7.6.1
django-debug-toolbar==4.4.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-smart-selects==1.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
phonenumberslite==8.13.37
pillow==10.3.0
progressbar2==4.4.2
PyJWT==2.8.0
python-utils==3.8.2
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.1
Unidecode==1.3.8
